Shot in the dark at why deposit work job is causing #1783
No idea what I am doing here at the moment. Finding it difficult to reason about this job which seems like a crucial one to understand. Extracted a service from the job to at least take job mechanics out of it.
